refactor: bls24315 hints
