+++ std/algebra/native/fields_bls24315/e12.go
@@ -17,10 +17,7 @@ limitations under the License.
 package fields_bls24315
 
 import (
-	"math/big"
-
 	bls24315 "github.com/consensys/gnark-crypto/ecc/bls24-315"
-	"github.com/consensys/gnark/constraint/solver"
 	"github.com/consensys/gnark/frontend"
 )
 
@@ -174,44 +171,6 @@ func (e *E12) Square(api frontend.API, x E12) *E12 {
 	return e
 }
 
-var InverseE12Hint = func(_ *big.Int, inputs []*big.Int, res []*big.Int) error {
-	var a, c bls24315.E12
-
-	a.C0.B0.A0.SetBigInt(inputs[0])
-	a.C0.B0.A1.SetBigInt(inputs[1])
-	a.C0.B1.A0.SetBigInt(inputs[2])
-	a.C0.B1.A1.SetBigInt(inputs[3])
-	a.C1.B0.A0.SetBigInt(inputs[4])
-	a.C1.B0.A1.SetBigInt(inputs[5])
-	a.C1.B1.A0.SetBigInt(inputs[6])
-	a.C1.B1.A1.SetBigInt(inputs[7])
-	a.C2.B0.A0.SetBigInt(inputs[8])
-	a.C2.B0.A1.SetBigInt(inputs[9])
-	a.C2.B1.A0.SetBigInt(inputs[10])
-	a.C2.B1.A1.SetBigInt(inputs[11])
-
-	c.Inverse(&a)
-
-	c.C0.B0.A0.BigInt(res[0])
-	c.C0.B0.A1.BigInt(res[1])
-	c.C0.B1.A0.BigInt(res[2])
-	c.C0.B1.A1.BigInt(res[3])
-	c.C1.B0.A0.BigInt(res[4])
-	c.C1.B0.A1.BigInt(res[5])
-	c.C1.B1.A0.BigInt(res[6])
-	c.C1.B1.A1.BigInt(res[7])
-	c.C2.B0.A0.BigInt(res[8])
-	c.C2.B0.A1.BigInt(res[9])
-	c.C2.B1.A0.BigInt(res[10])
-	c.C2.B1.A1.BigInt(res[11])
-
-	return nil
-}
-
-func init() {
-	solver.RegisterHint(InverseE12Hint)
-}
-
 // Inverse e12 elmts
 func (e *E12) Inverse(api frontend.API, e1 E12) *E12 {
 
@@ -232,57 +191,6 @@ func (e *E12) Inverse(api frontend.API, e1 E12) *E12 {
 	e.assign(res[:12])
 
 	return e
-}
-
-var DivE12Hint = func(_ *big.Int, inputs []*big.Int, res []*big.Int) error {
-	var a, b, c bls24315.E12
-
-	a.C0.B0.A0.SetBigInt(inputs[0])
-	a.C0.B0.A1.SetBigInt(inputs[1])
-	a.C0.B1.A0.SetBigInt(inputs[2])
-	a.C0.B1.A1.SetBigInt(inputs[3])
-	a.C1.B0.A0.SetBigInt(inputs[4])
-	a.C1.B0.A1.SetBigInt(inputs[5])
-	a.C1.B1.A0.SetBigInt(inputs[6])
-	a.C1.B1.A1.SetBigInt(inputs[7])
-	a.C2.B0.A0.SetBigInt(inputs[8])
-	a.C2.B0.A1.SetBigInt(inputs[9])
-	a.C2.B1.A0.SetBigInt(inputs[10])
-	a.C2.B1.A1.SetBigInt(inputs[11])
-
-	b.C0.B0.A0.SetBigInt(inputs[12])
-	b.C0.B0.A1.SetBigInt(inputs[13])
-	b.C0.B1.A0.SetBigInt(inputs[14])
-	b.C0.B1.A1.SetBigInt(inputs[15])
-	b.C1.B0.A0.SetBigInt(inputs[16])
-	b.C1.B0.A1.SetBigInt(inputs[17])
-	b.C1.B1.A0.SetBigInt(inputs[18])
-	b.C1.B1.A1.SetBigInt(inputs[19])
-	b.C2.B0.A0.SetBigInt(inputs[20])
-	b.C2.B0.A1.SetBigInt(inputs[21])
-	b.C2.B1.A0.SetBigInt(inputs[22])
-	b.C2.B1.A1.SetBigInt(inputs[23])
-
-	c.Inverse(&b).Mul(&c, &a)
-
-	c.C0.B0.A0.BigInt(res[0])
-	c.C0.B0.A1.BigInt(res[1])
-	c.C0.B1.A0.BigInt(res[2])
-	c.C0.B1.A1.BigInt(res[3])
-	c.C1.B0.A0.BigInt(res[4])
-	c.C1.B0.A1.BigInt(res[5])
-	c.C1.B1.A0.BigInt(res[6])
-	c.C1.B1.A1.BigInt(res[7])
-	c.C2.B0.A0.BigInt(res[8])
-	c.C2.B0.A1.BigInt(res[9])
-	c.C2.B1.A0.BigInt(res[10])
-	c.C2.B1.A1.BigInt(res[11])
-
-	return nil
-}
-
-func init() {
-	solver.RegisterHint(DivE12Hint)
 }
 
 // DivUnchecked e12 elmts
