+++ std/algebra/native/fields_bls24315/e4.go
@@ -17,10 +17,7 @@ limitations under the License.
 package fields_bls24315
 
 import (
-	"math/big"
-
 	bls24315 "github.com/consensys/gnark-crypto/ecc/bls24-315"
-	"github.com/consensys/gnark/constraint/solver"
 	"github.com/consensys/gnark/frontend"
 )
 
@@ -147,32 +144,6 @@ func (e *E4) Conjugate(api frontend.API, e1 E4) *E4 {
 	return e
 }
 
-var DivE4Hint = func(_ *big.Int, inputs []*big.Int, res []*big.Int) error {
-	var a, b, c bls24315.E4
-
-	a.B0.A0.SetBigInt(inputs[0])
-	a.B0.A1.SetBigInt(inputs[1])
-	a.B1.A0.SetBigInt(inputs[2])
-	a.B1.A1.SetBigInt(inputs[3])
-	b.B0.A0.SetBigInt(inputs[4])
-	b.B0.A1.SetBigInt(inputs[5])
-	b.B1.A0.SetBigInt(inputs[6])
-	b.B1.A1.SetBigInt(inputs[7])
-
-	c.Inverse(&b).Mul(&c, &a)
-
-	c.B0.A0.BigInt(res[0])
-	c.B0.A1.BigInt(res[1])
-	c.B1.A0.BigInt(res[2])
-	c.B1.A1.BigInt(res[3])
-
-	return nil
-}
-
-func init() {
-	solver.RegisterHint(DivE4Hint)
-}
-
 // DivUnchecked e4 elmts
 func (e *E4) DivUnchecked(api frontend.API, e1, e2 E4) *E4 {
 
@@ -192,28 +163,6 @@ func (e *E4) DivUnchecked(api frontend.API, e1, e2 E4) *E4 {
 	e.assign(res[:4])
 
 	return e
-}
-
-var InverseE4Hint = func(_ *big.Int, inputs []*big.Int, res []*big.Int) error {
-	var a, c bls24315.E4
-
-	a.B0.A0.SetBigInt(inputs[0])
-	a.B0.A1.SetBigInt(inputs[1])
-	a.B1.A0.SetBigInt(inputs[2])
-	a.B1.A1.SetBigInt(inputs[3])
-
-	c.Inverse(&a)
-
-	c.B0.A0.BigInt(res[0])
-	c.B0.A1.BigInt(res[1])
-	c.B1.A0.BigInt(res[2])
-	c.B1.A1.BigInt(res[3])
-
-	return nil
-}
-
-func init() {
-	solver.RegisterHint(InverseE4Hint)
 }
 
 // Inverse e4 elmts
